requirement.txt
pandas
numpy 
openai
python-dotenv
chromadb